Add support for deleting stored POI
It is now possible to delete stored POI from the POI database
by pressing the "Delete POI from database" button on a POI detail
page & confirming the decision in a follow-up "are you sure" page.

This is deliberately not a top menu action as there are already
three actions in the top menu and it is expected POIs will be deleted rather seldom. So this way the top menu is not cluttered and users don't have to worry about selecting a "Delete" option by mistake (there is a confirmation dialog but users might not be aware of that).

When deleting POI we also make sure to delete any corresponding markers from the map and to reload the category listing.