Tweak TwoOptionPage layout and button colors
Make sure the buttons have sane background color (themed rectangle, which is the base for the text button, is by default transparent on Sailfish OS). Also center the message appropriately on mobile devices & desktop and increase the spacing between the buttons to make them more distinct and to make pressing the wrong button less likely.