PEP 8 + properties for device modules & paths module
Apply PEP8 and turn relevant methods to properties for
all device modules and also for the paths modules which
is pretty interconnected with the device modules.

This was actually not trivial due to the need to carefully
checks all the refactoring was applied correctly and properties
are used instead of method calls everywhere a method turned into
a property.

In general modRana seems to be working fine & the few tests we have
are passing, so hopefully no nasty surprises will show up as
a result of this refactoring.