Don't version when loading GPX files
Modern upoints library no longer supports passing a GPX version
when parsing GPX files, so tracklog loading might fail if a modern
upoints library is installed on a system where modRana is running,
as it apparently takes precedence before the old bundled one.

So drop the version requirement, as old upoints also seems
to work fine without it. This should make modRana compatible
with both the new and old version.