Remove the old Qt4/QML GUI
This was the first non GTK 2 UI started as a GUI that would run
on the N9/N950 that never supported GTK 2 good enough.

After a lot of mucking, it was also possible to get this UI running
on the N900 as well.

The UI never really got proper feature parrity with the "classic"
GTK 2 UI and development more or less ceased when overall N9 development
ceased and the buggy N900 Qt 4.7/QML libraries were never fixed.

So it's time to remove this part of the codebase, but it will live
forever in our git history. :)