Handle native launchers in QML argv processing
The modRana QML does some argv processing on its side
before it is passed to the modRana Python backend due to
some unwanted arguments QML launchers such as qmlscene
inject to argv.

This type of processing is not needed for native launchers
so skip it if QML detects it has been started by a native launcher.