Fix native launcher detections
Of course, the name based native launcher detection turned out to be
fragile as when the native launcher is invoked from from the icon, it
gets the full path to its binary, not binary name... :D

So lets to the right thing, drop the binary name based native launcher
detection and have the native launcher set a property on the root
QML context.

Also fix the apparently previously inconsistent native launcher C++
code indentation.