Remove the GTK2 GUI
After 10 years it's time to say good by to the classical modRana GTK2 GUI. This GUI started as the original Rana GUI for the Neo FreeRunner and used a custom Cairo based GUI toolkit, as back then GTK was not really suitable for dynamic graphical interfaces on mobile device.

This enabled extreme control over what was drawn, yet had quite some overhead as all elements were custom drawn and implemented. It was also not really possible to implement animations (other than the hyper optimized custom map dragging animation), so stuff like kinetic scrolling in menus was not really possible.

The next GUI modRana got was a Qt 4 + QtQuick 1.0 GUI targeting MeeGo Harmattan on the N9/950. This GUI was pretty nice, but never reached feature parity with the GTK UI, mostly due to the original MeeGo devices not having any direct successors.

The current modern Qt 5 + QtQuick 2.0 GUI originated from the effort to port modRana to Sailfish OS, starting a rather long effort to provide feature parity with the GTK GUI. This was finally reached when the modRana Qt 5 GUI could finally provide routing, PoI handling and track logging/display.

Now back to the GTK2 GUI. It's 2020 and the number of modRana users on the N900 is most likely very low (and we have a branch for N900 if we really need to do a new release :D). GTK2 is Python 2 only and Python 2 went EoL on January 1 2020. So it's time to say goodbye to the classical modRana GTK2 GUI and thanks for all the fish! :)