Remove the horrible Python 2.5 property setter hack
So in Python 2.5 properties were read only didn't yet have a setter, yet we had to support Python 2.5 due to this being the only usable Python version on the N900.

Now that we are Python 3 only & have a special branch for any N900 changes or release, we can finally remove this horrible hack. :)